Fix a bug that caused parsed_to to be set to null
When the last commit from the Beanstalk API was a subversion commit then
the hash_id in the last commit would be empty so PHP would try to save
a null instead.

Prevent this by checking if hash_id exists on the object and only saving
it to the config if it does. This works because the vast majority of our
commits is done using git so the next time the script runs it will
update the config.